Add setup for runtime execution control
Originally implemented a with a C++ JSON parser, but it was a shitshow:
    - how do we pass errors back to Go?
    - the parser was handling everything by throwing
    - we'll still need to parse from Go later when 'debugging' mode is
      implemented
So we changed our mind, and will implement the parsing in Go entirely.
Originally we wanted to avoid converting values from Go to C to C++, but
that seems better than having 2 parsers, even with all the drawbacks.