Lock OCR execution thread to OS thread
As suggested in a Reddit thread, this reduces Go scheduler conflicts
which happen when C calls are long (the thread states 1μs or more).
Since all our calls are >1ms, this significantly improves performance:
from ~26ms down to ~16ms on the 'neograf' test set (average total OCR
pipeline execution).

Note however, that the thread also states that when doing this, we
should avoid parking the (OS-locked) executing Go routine, since this
will cause a context switch and a non-negligible overhead (several ms)
when waking the Go routine.