Implement PTP
Wanted to use this for synchronizing the acquired image time stamp with
the host machine system clock, primarily so that the image acquisition
and transfer time can be computed.
However, this is fairly unstable, since the image time stamp can be
lower than the received time stamp due to the clock offsets. Considering
that the offset can only be determined for some interval in an average
sense, this doesn't seem like a good method for computing the transfer
time.

Hence, this method won't be used. PTP can still be enabled, however, through
runtime selectable parameters, we simply won't use it explicitly.

The acquisition/transfer time will simply be limited using timeouts.
The image will be timestamped once received, however we'll implement
this in Go code, not in C++ (libcamera).

NOTE: for PTP to work properly, ptp4l needs to be running on the host
machine (or some other clock master) or a PTP clock needs to be
available on the network.
When using ptp4l, 'hardware' stamping should be used, since 'software'
stamping is fairly inaccurate and unstable (the offset oscillates a
lot). When using 'hardware' stamping, 'phc2sys' also needs to be running
on the host machine for the camera to properly set it's own time stamp.
Hence the following should be run on the host machine, before
initializing the camera PTP:

	$ ptp4l -i <interface (eg. eth0)> -E --priority1=127
	$ phc2sys -a -r -r

In this configuration, the camera will set it's time stamp to the host
machine system clock.