Reimplement C-memory management helper type
Handling arrays of C-structs with pointers was very impractical with the
previous implementation.

Since a reallocation could occur during any memory request, filling an
array of C-structs which contain pointers would require first allocating
and storing all pointers and values, and then filling the array, i.e. we
would have to loop over the array twice, since the C-struct pointers
could get invalidated because of a reallocation.
We also can't ensure a reallocation won't take place, e.g. by allocating
enough memory initially, since we mostly want to use the manager with
data defined at runtime (configs for 'Init()' functions).

Apart from this, storing the locations of pointers for an arbitrary
struct is also super impractical: we either have to create a dedicated
type, which would map the locations to the struct; create an anonymous
type, which looks horrible; using a bunch of slices, one for each field;
or adjusting the C-shim layer to work with a base pointer and an offset,
which is the worst of the three options.

So, we've reimplemented the manager to just use non-contiguous memory
because this makes things much simpler, and is essentially how we've
been doing it up to this point, but we've just been doing the 'mallocs'
and 'defer frees' manually.
If we come up with an easy-to-use solution for the contiguous memory
manager, we'll reimplement it.