Major code refactor
- handle C-string arrays with mem.Arena
- fix some memory leaks
- add a Go convenience type for handling timings (chrono.Timing)
- split image handling, OCR and output into separate packages
- expand OCR results
	- return per-line detection/recognition results from C
	- join per-line results to a coherent structure in Go
	- add per-line ROI for the recognised text
	- add per-line confidences for the recognised text
- update beholder subcommands
- bump tesseract (5.3.4 -> 5.4.1) and OpenCV (4.9.0 -> 4.10.0) versions
- update third party C-libs build script

TODO:
- further decouple Go packages, e.g. 'camera' from 'image'
- implement a dedicated Image type in Go and refactor C code so that the
  image processor doesn't own the image
- refactor the 'output' package
- C logging and error handling

NOTE:
There has been a severe performance degradation of the text recognition
process: on averaged, from ~15ms to ~30ms on the 'neograf' test set.
Based on manual timing and review, neither the C-Go interface layer or
the Go code itself seem to be the issue --- the (third party) C
functions themselves seem to have suffered the degradation.

Both tesseract and OpenCV functions are taking longer to execute, even
when rolling back 'libbeholder' changes (which were fairly small,
especially performance-wise, in the grand scheme of things).

There also seems to be an initial cost when calling
Processor::preprocess() for the first time, which is about the regular
execution time of subsequent calls, i.e. the first call takes twice as
long as subsequent ones. It is not clear as to why this is happening at
this point, or whether this was happening before --- image processing
times were significantly shorter before (~3ms on the 'neograf' set on
average), hence this might have been happening all along.

At this point, we've narrowed it down to two possible issues:

	1. Switching from a static 'libbeholder' build to a dynamic
	   library (and linkage): when we updated we forgot to
	   benchmark.
	   We'll test this by rolling back to a commit in which we know
	   the performance (02f0a99) and rebuild everything with the
	   current setup (GCC 12, Debian 12)
	2. GCC (and/or kernel) issues: we've updated both GCC (11 -> 12),
	   and Debian (11 -> 12), hence this might be causing issues.
	   We'll test GCC with a Docker container; we'll have to think
	   of something for testing the impact of the kernel.

For future reference, this is the current GCC and kernel info:

    - Linux ph-AMD 6.1.0-23-amd64 #1 SMP PREEMPT_DYNAMIC Debian 6.1.99-1 (2024-07-15) x86_64 GNU/Linux
    - gcc (Debian 12.2.0-14) 12.2.0

These are the average OCR timings for the 'neograf' set:

	read           : 4.5626ms
	decode         : 2.755328ms
	preprocess     : 7.562029ms
	ocr            : 22.985681ms
	postprocess    : 1.495µs
	output         : 134.148µs
	total          : 38.001617ms

	init: 90.131006ms
	exec: 812.635153ms

This is a sample of OCR timings for a single image from the 'neograf'
set (imagefile_14.bmp):

	read           : 3.552458ms
	decode         : 1.288846ms
	preprocess     : 21.262836ms
	ocr            : 23.717999ms
	postprocess    : 2.3µs
	output         : 343.452µs
	total          : 50.168091ms

	init: 89.89576ms
	exec: 140.080161ms