Make C-API build work without a third-party package (archive)
libbeholder can now be built without providing a third-party
archive/MD5, i.e., the third-party dependencies can be installed
manually, globally or into a staging directory, and the build will
correctly set include/link directories.

The purpose of this change is to simplify top-level build scripts, which
might want to use a staging directory, but don't necessarily want to
hassle with the third-party archive.