Track tmpfs-mounting implementation
Not used (or tested) at this point.
Mounting might cause complications (e.g. in Docker containers, special
permissions, etc.), and the benefits are minimal for our current use
case --- the uncompressed embeds are about ~30MB in total, and the
output will happen at program startup, so it shouldn't be noticeable.

Should remove entirely if we don't end up using it.